cohort 1, chapters 1-3 (#3)
* cohort 1, chapters 1-3

* made slides
* edited "club intro" to link to our textbook
* added "images" as subdirectory of "slides"

* Remove extra title slide.

The "1. Preliminaries" title (in the header yaml) is what shows up in the index, but it also automatically acts as the first slide.

I also turned off incremental bulleted lists by default, because I find myself manually turning them off most of the time.

I tried to fix the styling so you wouldn't have to insert the empty spacer columns, because those are a clear sign that something is wrong (you should be able to just put content in your qmd and let styling deal with formatting), but the revealjs style fought me. I might come back to it later.

* Remove extraneous session info block.

* I don't think these callout notes are displaying the way you intend. Did you preview the slides? I feel like you must have because we discussed them, but what I'm seeing looks broken.

---------

Co-authored-by: Jon Harmon <jonthegeek@gmail.com>